# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Haraj One API
  version: 2.0.0
paths:
  /api/v1/auctions/:
    get:
      operationId: auctions_list
      description: '`GET /api/v1/auctions/` — the auctions a caller may see, newest
        first.'
      summary: قائمة المزادات
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          default: 0
      - in: query
        name: state
        schema:
          enum:
          - draft
          - scheduled
          - live
          - ended
          - settled
          - cancelled
          - ''
          type: string
        description: |-
          * `draft` - مسودة
          * `scheduled` - مجدول
          * `live` - جارٍ
          * `ended` - منتهٍ
          * `settled` - مُسوّى
          * `cancelled` - ملغى
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionPage'
          description: ''
  /api/v1/auctions/{id}/:
    get:
      operationId: auctions_retrieve
      description: |-
        `GET /api/v1/auctions/{id}/` — one auction, in the list's own shape.

        Built by `cards.auction_card` with the same annotations the list uses, so
        the detail page and the row a customer tapped cannot disagree about how many
        cars an auction holds. An auction the caller may not see is a 404: a 403
        would confirm it exists, which is enough to enumerate auctions before they
        open.
      summary: تفاصيل المزاد
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuctionCard'
          description: ''
  /api/v1/auctions/{id}/vehicles/:
    get:
      operationId: auctions_vehicles_list
      description: '`GET /api/v1/auctions/{id}/vehicles/` — one auction''s cars.'
      summary: مركبات المزاد
      parameters:
      - in: query
        name: auction
        schema:
          type: integer
          minimum: 1
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - in: query
        name: limit
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
      - in: query
        name: make
        schema:
          type: string
          maxLength: 80
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          default: 0
      - in: query
        name: search
        schema:
          type: string
          maxLength: 80
      - in: query
        name: state
        schema:
          enum:
          - draft
          - listed
          - bidding
          - awaiting_decision
          - awarded
          - rejected
          - invoiced
          - paid
          - released
          - withdrawn
          - relisted
          - ''
          type: string
        description: |-
          * `draft` - مسودة
          * `listed` - معروضة
          * `bidding` - تحت المزايدة
          * `awaiting_decision` - بانتظار قرار المالك
          * `awarded` - مرسّاة
          * `rejected` - مرفوضة
          * `invoiced` - مفوترة
          * `paid` - مسدَّدة
          * `released` - خرجت
          * `withdrawn` - مسحوبة
          * `relisted` - معادة للعرض
      - in: query
        name: year_from
        schema:
          type: integer
          maximum: 2200
          minimum: 1900
      - in: query
        name: year_to
        schema:
          type: integer
          maximum: 2200
          minimum: 1900
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiclePage'
          description: ''
...
  /api/v1/favourites/:
    get:
      operationId: favourites_list
      description: |-
        `GET /api/v1/favourites/` — the cars this customer marked, newest first.

        Rendered through `cards.vehicle_card`, like every other list of vehicles in
        the product: a favourites screen that assembled its own row would be the
        second card builder `ops/checks/one_vehicle_card.py` exists to refuse, and
        the field that went missing from it would be missing only here.

        The visibility rule still applies. A car marked while it was listed and
        since withdrawn is not shown — a favourite is a bookmark, never a claim, and
        it does not grant sight of a row its owner may no longer see.
      summary: المفضّلة
      parameters:
      - in: query
        name: limit
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          default: 0
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiclePage'
          description: ''
  /api/v1/favourites/{id}/:
    put:
      operationId: favourites_mark
      description: |-
        `PUT` and `DELETE /api/v1/favourites/{id}/` — mark and unmark.

        `PUT`, not `POST`, and both are idempotent: marking twice is marking once,
        and unmarking something unmarked is not an error. That is what a
        double-tapped heart and a retried request produce, and «هذه المركبة في
        مفضّلتك بالفعل» is a refusal for a thing that already happened the way the
        customer wanted.

        Both answer `204`. There is nothing to return — the client already knows
        which car it asked about, and a body here would be a second place the mark's
        shape is described.
      summary: إضافة إلى المفضّلة
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
    delete:
      operationId: favourites_unmark
      description: |-
        `PUT` and `DELETE /api/v1/favourites/{id}/` — mark and unmark.

        `PUT`, not `POST`, and both are idempotent: marking twice is marking once,
        and unmarking something unmarked is not an error. That is what a
        double-tapped heart and a retried request produce, and «هذه المركبة في
        مفضّلتك بالفعل» is a refusal for a thing that already happened the way the
        customer wanted.

        Both answer `204`. There is nothing to return — the client already knows
        which car it asked about, and a body here would be a second place the mark's
        shape is described.
      summary: إزالة من المفضّلة
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/vehicles/:
    get:
      operationId: vehicles_list
      description: '`GET /api/v1/vehicles/` — cars across auctions, searched and filtered.'
      summary: قائمة المركبات
      parameters:
      - in: query
        name: auction
        schema:
          type: integer
          minimum: 1
      - in: query
        name: limit
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
      - in: query
        name: make
        schema:
          type: string
          maxLength: 80
      - in: query
        name: offset
        schema:
          type: integer
          minimum: 0
          default: 0
      - in: query
        name: search
        schema:
          type: string
          maxLength: 80
      - in: query
        name: state
        schema:
          enum:
          - draft
          - listed
          - bidding
          - awaiting_decision
          - awarded
          - rejected
          - invoiced
          - paid
          - released
          - withdrawn
          - relisted
          - ''
          type: string
        description: |-
          * `draft` - مسودة
          * `listed` - معروضة
          * `bidding` - تحت المزايدة
          * `awaiting_decision` - بانتظار قرار المالك
          * `awarded` - مرسّاة
          * `rejected` - مرفوضة
          * `invoiced` - مفوترة
          * `paid` - مسدَّدة
          * `released` - خرجت
          * `withdrawn` - مسحوبة
          * `relisted` - معادة للعرض
      - in: query
        name: year_from
        schema:
          type: integer
          maximum: 2200
          minimum: 1900
      - in: query
        name: year_to
        schema:
          type: integer
          maximum: 2200
          minimum: 1900
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehiclePage'
          description: ''
  /api/v1/vehicles/{id}/:
    get:
      operationId: vehicles_retrieve
      description: |-
        `GET /api/v1/vehicles/{id}/` — one car, through the one card builder.

        T609's whole requirement is that this returns the *same* fields as the list.
        It does so by construction rather than by discipline: both call
        `cards.vehicle_card`, and there is no second assembly of a vehicle anywhere
        (`ops/checks/one_vehicle_card.py` fails the build if one appears).

        A car the caller may not see is a 404, not a 403. A 403 confirms the row
        exists, which is enough to enumerate an auction before it opens.
      summary: تفاصيل المركبة
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - bearerAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleCard'
          description: ''
...
components:
  schemas:
    AuctionCard:
      type: object
      description: The auction row, as `cards.auction_card` builds it.
      properties:
        id:
          type: integer
        number:
          type: integer
        title:
          type: string
        state:
          type: string
        state_label:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        vehicle_count:
          type: integer
          nullable: true
        open_vehicle_count:
          type: integer
          nullable: true
      required:
      - ends_at
      - id
      - number
      - open_vehicle_count
      - starts_at
      - state
      - state_label
      - title
      - vehicle_count
    AuctionPage:
      type: object
      properties:
        total:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/AuctionCard'
      required:
      - results
      - total
...
    VehicleCard:
      type: object
      description: |-
        The vehicle card — the same fields in the list and on the detail page.

        T609's acceptance criterion is that those two are identical, and
        `test_vehicle_api.py` asserts it by comparing the key sets rather than by
        reading this class, so the guarantee does not depend on this file being
        kept honest by hand.
      properties:
        id:
          type: integer
        auction_number:
          type: integer
        auction_state:
          type: string
        lot_number:
          type: integer
        title:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        odometer_km:
          type: integer
          nullable: true
        transmission:
          type: string
        transmission_label:
          type: string
        fuel_type:
          type: string
        fuel_type_label:
          type: string
        condition:
          type: string
        condition_label:
          type: string
        plate_type:
          type: string
        plate_type_label:
          type: string
        reserve_price:
          type: string
          nullable: true
        state:
          type: string
        state_label:
          type: string
        listing_state:
          type: string
        owner_company_name:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
      required:
      - auction_number
      - auction_state
      - condition
      - condition_label
      - fuel_type
      - fuel_type_label
      - id
      - listing_state
      - lot_number
      - make
      - model
      - odometer_km
      - owner_company_name
      - plate_type
      - plate_type_label
      - reserve_price
      - state
      - state_label
      - thumbnail_url
      - title
      - transmission
      - transmission_label
      - year
    VehiclePage:
      type: object
      properties:
        total:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/VehicleCard'
      required:
      - results
      - total
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: رمز الوصول من `/api/v1/auth/verify/`. صالح لخمس عشرة دقيقة، ويُجدَّد
        من `/api/v1/auth/refresh/`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid